requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
finnhub-python==2.4.23
idna==3.10
kafka-python==2.1.4
numpy==2.2.4
pandas==2.2.3
psycopg2-binary==2.9.10
pyarrow==19.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
six==1.17.0
tzdata==2025.2
urllib3==2.3.0
